+++ src/main/java/org/sopt/app/application/auth/dto/RecommendFriendRequest.java
@@ -1,7 +1,7 @@
 package org.sopt.app.application.auth.dto;
 
 import java.util.List;
-import org.sopt.app.application.auth.dto.PlaygroundProfileInfo.RecommendFriendFilter;
+import org.sopt.app.application.auth.dto.RecommendedFriendInfo.RecommendFriendFilter;
 
 public record RecommendFriendRequest(
         List<Integer> generations,
